refactor(v0.6): split core/wiki_generator/_frontmatter.py 21.4KB → 4-file package
CLAUDE.md rule #5: no file in core/ exceeds 20 KB. Legacy single
file (21.4 KB) sits above the cap. Split via MRO sub-mixin
composition preserves the public surface byte-identically.

Sub-files (all < 20 KB):
  __init__.py      2.3 KB   composes WikiFrontmatterMixin from 4
                            sub-mixins via MRO (no method-name
                            collisions; order is documentation-driven)
  init_state.py    3.0 KB   WikiInitStateMixin: __init__ +
                            _create_index_template (holds the
                            late-binding WIKI_DIR import the test
                            monkey-patch pattern depends on)
  id_gen.py        0.9 KB   WikiIdGenMixin: _generate_entity_id +
                            _normalize_name
  read.py          5.7 KB   WikiReadMixin: index build, overlap
                            snapshot, search, frontmatter read,
                            sensitivity default, duplicate detect
  create.py       12.3 KB   WikiCreateMixin: create_entity_file
                            (the orchestrator that dominates)

Refactor 5/5 (final) of the v0.6 oversize-module cleanup. The full
20 KB cap is now satisfied across core/ — and three lock-tests
(reflect, gemma_client, replay_graph, ingestion, frontmatter)
prevent regrowth past the cap on future PRs.

Quality delta: exempt (label: chore) — pure mechanical split, no
behaviour change. All canonical methods preserved on
WikiFrontmatterMixin via MRO; WikiGenerator(source_type="test")
end-to-end construction smoke verified.

Regression sweep stays green: test_v06_frontmatter (8/8 new),
test_phase_b (13/13), test_event (62/62),
test_attributes_summary_cleanup + test_entity_type_extension
(13/13).

Closes #N/A (CLAUDE.md rule #5 hygiene — refactor 5/5 / FINAL).
